Move logger helpers to pkg/util
